fix: build_terminal_modesのcfgetispeed/cfgetospeedがmacOSでu64→u32型不一致になっていたのを修正
libc::speed_tはLinuxではc_uint(u32)だがmacOSではc_ulong(u64)で、u32を
要求するrussh::Ptyのタプルにそのまま渡すとmacOSビルドのみ型不一致で
コンパイルエラーになっていた（このセッションの変更とは無関係、native
pty gapsブランチに元から存在した既存バグ。test-macos CIが本セッション
以前のコミットから既に失敗していたことを確認済み）。実際のボーレート値が
u32::MAXへ達することはないため`as u32`で安全にキャストする。

+++ rust-core/isekai-ssh/src/native/console.rs
@@ -297,9 +297,11 @@ pub(crate) fn build_terminal_modes() -> Vec<(russh::Pty, u32)> {
                 push_oflag!(ONOCR, ONOCR);
                 push_oflag!(ONLRET, ONLRET);
 
-                // Speed (baud rate)
-                let ispeed = libc::cfgetispeed(&termios);
-                let ospeed = libc::cfgetospeed(&termios);
+                // Speed (baud rate). `libc::speed_t` is `u32` on Linux but
+                // `u64` on macOS — `as u32` is lossless in practice (real
+                // baud rates never approach `u32::MAX`).
+                let ispeed = libc::cfgetispeed(&termios) as u32;
+                let ospeed = libc::cfgetospeed(&termios) as u32;
                 modes.push((russh::Pty::TTY_OP_ISPEED, ispeed));
                 modes.push((russh::Pty::TTY_OP_OSPEED, ospeed));
                 return modes;
